requirements.txt
PySide6
python-mpv
pyinstaller
mutagen
